backend/requirements.txt
# Django and REST framework dependencies
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.1.0
python-dotenv>=1.0.0

# Authentication
django-allauth>=0.54.0
dj-rest-auth>=4.0.1

# DOCX processing
docx2txt>=0.8
python-docx>=1.0.0

# Utilities
pillow>=10.0.0

# Ollama integration
numpy>=1.24.3
httpx>=0.24.1
requests>=2.31.0

# Anthropic (Claude) API
anthropic>=0.8.0 